backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
email-validator
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
python-dateutil
pdfplumber
